requirements.txt
amqp==5.3.1
asgiref==3.9.2
billiard==4.2.2
celery==5.5.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.6
django-cors-headers==4.9.0
django-crontab==0.7.1
django-filter==25.1
django-seed==0.3.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
Faker==37.8.0
gunicorn==23.0.0
inflection==0.5.1
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
six==1.17.0
sqlparse==0.5.3
toposort==1.10
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.14
